requirement.txt
albumentations
braceexpand
chardet
ftfy
h5py
huggingface_hub
imbalanced_learn
joblib
lightning
nltk
opencv_python
pandas
Pillow
pytorch_lightning
PyYAML
regex==2024.7.24
scikit_learn
setuptools
stanza
submitit
timm
torchmetrics
tqdm
transformers
webdataset
xgboost
